rag-eval-harness/requirements.txt
pandas>=1.4.0
numpy>=1.20.0
tqdm>=4.64.0
pyyaml>=6.0
python-dotenv>=1.0.0
rich>=13.4.0
requests>=2.28.0
openai>=1.0.0
